examples/python/requirements.txt
async-timeout==4.0.2
Deprecated==1.2.13
packaging==21.3
pyparsing==3.0.9
python-decouple==3.6
redis==4.3.4
wrapt==1.14.1
